implemented function to read config in utils
