+++ src/train.py
@@ -1,3 +1,4 @@
+import os
 import sys
 import time
 from pathlib import Path
@@ -10,28 +11,18 @@ import torchvision.transforms as transforms
 from tqdm import tqdm
 from torch.utils.tensorboard import SummaryWriter
 from pytorch_gan_metrics import get_inception_score_and_fid
-import yaml
 
 import nets
-
-from utils import gradient_penalty, initialize_weights
-
 import losses
-
-import os
-
-os.environ["OMP_NUM_THREADS"] = "4"
+from utils import gradient_penalty, initialize_weights, read_config
 
 # for windows
 import ssl
 
 ssl._create_default_https_context = ssl._create_unverified_context
 
+os.environ["OMP_NUM_THREADS"] = "4"
 working_dir = Path(__file__).parent.parent.absolute()
-unique_key = str(str(time.ctime())).replace(" ", "_").replace(":", "_")
-experiments_dir = Path()  # set this paths after reading the config file
-gen_dir = Path()
-disc_dir = Path()
 
 
 def main():
@@ -60,14 +51,17 @@ def main():
         LAMBDA_GP,
         GEN_LOSS_STR,
         DISC_LOSS_STR,
+        EXPERIMENT_DIR,
+        GEN_DIR,
+        DISC_DIR,
     ) = read_config(sys.argv)
 
-    Path(experiments_dir).mkdir(parents=True, exist_ok=True)
-    Path(gen_dir.parent).mkdir(parents=True, exist_ok=True)
-    open(gen_dir, "w+")
-    open(disc_dir, "w+")
-    print("Saving experiment under: \t", experiments_dir)
-    print("Saving experiment models under: ", gen_dir.parent)
+    Path(EXPERIMENT_DIR).mkdir(parents=True, exist_ok=True)
+    Path(GEN_DIR.parent).mkdir(parents=True, exist_ok=True)
+    open(GEN_DIR, "w+")
+    open(DISC_DIR, "w+")
+    print("Saving experiment under: \t", EXPERIMENT_DIR)
+    print("Saving experiment models under: ", GEN_DIR.parent)
 
     np.random.seed(SEED)
     torch.manual_seed(SEED)
@@ -114,7 +108,7 @@ def main():
     initialize_weights(generator)
     initialize_weights(discriminator)
 
-    writer = SummaryWriter(experiments_dir)
+    writer = SummaryWriter(EXPERIMENT_DIR)
 
     gen_optimizer = torch.optim.Adam(
         generator.parameters(), lr=GEN_LR, betas=(0.9, 0.999)
@@ -152,9 +146,7 @@ def main():
 
                 gp = 0.0
                 if LAMBDA_GP != 0:
-                    gp = gradient_penalty(
-                        discriminator, labels, data, fake, device=device
-                    )
+                    gp = gradient_penalty(discriminator, labels, data, fake, device)
 
                 loss_disc = (loss_real + loss_fake) + LAMBDA_GP * gp
                 batch_loss_disc.append(loss_disc.item())
@@ -251,145 +243,15 @@ def main():
             writer.add_scalar("test_accuracy/fake", accuracy_fake, step)
             writer.add_scalar("evaluation/inception_score", incep_score, step)
             writer.add_scalar("evaluation/inception_std", incep_score_std, step)
-            writer.add_scalar("evaluation/test_frechet_distance", frechet_distance, step)
+            writer.add_scalar("evaluation/frechet_distance", frechet_distance, step)
             grid_real = torchvision.utils.make_grid(imgs_real, nrow=16, normalize=True)
             grid_fake = torchvision.utils.make_grid(imgs_fake, nrow=16, normalize=True)
             writer.add_image("real", grid_real, step)
             writer.add_image("fake", grid_fake, step)
 
         if step % SAVE_EVERY == 0:
-            torch.save(generator.state_dict(), gen_dir)
-            torch.save(discriminator.state_dict(), disc_dir)
-
-
-def read_config(_input):
-    if len(_input) == 1:
-        print(
-            "No configuration file specified.\n"
-            "Looking for default configuration file under configs/ directory."
-        )
-        path_config = working_dir / Path("configs/default.yaml")
-    else:
-        print("Train model using %s as configuration file." % (_input[1]))
-        path_config = working_dir / Path("{}".format(_input[1]))
-
-    global experiments_dir, gen_dir, disc_dir
-    experiments_dir = (
-        working_dir / Path("experiments/" + path_config.stem) / Path(unique_key)
-    )
-    gen_dir = (
-        working_dir
-        / Path("models/" + path_config.stem)
-        / Path(unique_key)
-        / Path("gen.pt")
-    )
-    disc_dir = (
-        working_dir
-        / Path("models/" + path_config.stem)
-        / Path(unique_key)
-        / Path("disc.pt")
-    )
-    if path_config.suffix != ".yaml":
-        print("Make sure that the configuration file is a .yaml file.")
-        sys.exit(1)
-
-    try:
-        with open(path_config) as yamlfile:
-            config = yaml.load(yamlfile, Loader=yaml.FullLoader)
-    except OSError as e:
-        print("Configuration file not available, check under configs/ directory.")
-        print(e)
-        sys.exit(1)
-
-    try:
-        config_dataset = (
-            img_size,
-            channels_img,
-            num_classes,
-        ) = list(config["dataset"].values())
-
-        config_model = (
-            generator_model,
-            discriminator_model,
-            disc_features,
-            gen_features,
-            latent_dim,
-            embedding_dim,
-        ) = list(config["nets"].values())
-
-        config_training = (
-            batch_size,
-            test_batch_size,
-            test_every,
-            save_every,
-            epochs,
-            gen_lr,
-            disc_lr,
-            gamma,
-            cuda,
-            seed,
-            disc_iterations,
-            weight_clip,
-            lambda_gp,
-            gen_loss,
-            disc_loss,
-        ) = list(config["training"].values())
-
-        assert type(img_size) == int
-        assert type(channels_img) == int
-        assert type(num_classes) == int
-        assert type(generator_model) == str
-        assert type(discriminator_model) == str
-        assert type(disc_features) == int
-        assert type(gen_features) == int
-        assert type(latent_dim) == int
-        assert type(embedding_dim) == int
-        assert type(batch_size) == int
-        assert type(test_batch_size) == int
-        assert type(test_every) == int
-        assert type(save_every) == int
-        assert type(epochs) == int
-        assert type(gen_lr) == float
-        assert type(disc_lr) == float
-        assert type(gamma) == float
-        assert type(cuda) == bool
-        assert type(seed) == int
-        assert type(disc_iterations) == int
-        assert type(weight_clip) == float
-        assert type(lambda_gp) == int
-        assert type(gen_loss) == str
-        assert type(disc_loss) == str
-    except (AssertionError, ValueError, KeyError) as e:
-        print("The given .yaml file uses a wrong convention.")
-        print(
-            "The expected format for the .yaml file is:\n"
-            "dataset:\n"
-            "    img_size: int\n"
-            "    channels_img: int\n"
-            "    num_classes: int\n"
-            "nets:\n"
-            "    disc_features: int\n"
-            "    gen_features: int\n"
-            "    latent_dims: int\n"
-            "    embedding_dim: int\n"
-            "training:\n"
-            "    batch_size: int\n"
-            "    test_batch_size: int\n"
-            "    test_every: int\n"
-            "    save_every: int\n"
-            "    epochs: int\n"
-            "    gen_lr: float\n"
-            "    disc_lr: float\n"
-            "    gamma: float\n"
-            "    cuda: bool\n"
-            "    seed: int"
-            "    gen_loss: str"
-            "    disc_loss: str"
-        )
-        print(e)
-        sys.exit(1)
-
-    return config_dataset + config_model + config_training
+            torch.save(generator.state_dict(), GEN_DIR)
+            torch.save(discriminator.state_dict(), DISC_DIR)
 
 
 if __name__ == "__main__":
